Move hlint tests to its package
