+++ plugins/hls-hlint-plugin/hls-hlint-plugin.cabal
@@ -101,3 +101,20 @@ library
   default-extensions:
     DataKinds
     TypeOperators
+
+test-suite tests
+  type:             exitcode-stdio-1.0
+  default-language: Haskell2010
+  hs-source-dirs:   test
+  main-is:          Main.hs
+  ghc-options:      -threaded -rtsopts -with-rtsopts=-N
+  build-depends:
+      aeson
+    , base
+    , filepath
+    , hls-hlint-plugin
+    , hls-plugin-api
+    , hls-test-utils      >=1.0 && <1.2
+    , lens
+    , lsp-types
+    , text
