+++ plugins/hls-hlint-plugin/test/testdata/ApplyRefact2.hs
