+++ plugins/hls-hlint-plugin/test/testdata/ApplyRefact3.hs
