+++ plugins/hls-hlint-plugin/test/testdata/ApplyRefact4.hs
