+++ plugins/hls-hlint-plugin/test/testdata/ApplyRefact5.hs
