+++ plugins/hls-hlint-plugin/test/testdata/ApplyRefact6.hs
