+++ plugins/hls-hlint-plugin/test/testdata/ApplyRefact7.hs
