+++ plugins/hls-hlint-plugin/test/testdata/ApplyRefact8.hs
