+++ plugins/hls-hlint-plugin/test/testdata/cpp/ApplyRefact2.hs
