+++ plugins/hls-hlint-plugin/test/testdata/cpp/ApplyRefact3.hs
