+++ plugins/hls-hlint-plugin/test/testdata/lambdacase/ApplyRefact1.hs
