+++ plugins/hls-hlint-plugin/test/testdata/lambdacase/hie.yaml
