+++ plugins/hls-hlint-plugin/test/testdata/test-hlint-config.yaml
