+++ plugins/hls-hlint-plugin/test/testdata/typeapps/ApplyRefact1.hs
