+++ plugins/hls-hlint-plugin/test/testdata/typeapps/hie.yaml
